Refactor didiator files structure
