+++ didiator/dispatchers/__init__.py
@@ -1,0 +1,9 @@
+from .command import CommandDispatcherImpl
+from .request import RequestDispatcherImpl
+from .query import QueryDispatcherImpl
+
+__all__ = (
+    "CommandDispatcherImpl",
+    "RequestDispatcherImpl",
+    "QueryDispatcherImpl",
+)
