+++ didiator/dispatchers/command.py
@@ -1,10 +1,10 @@
 from typing import Any, Type, TypeVar
 
-from didiator.command import Command
-from didiator.interface.command_dispatcher import CommandDispatcher
-from didiator.interface.dispatcher import HandlerType
+from didiator.interface.entities.command import Command
+from didiator.interface.dispatchers.command import CommandDispatcher
+from didiator.interface.handlers.request import HandlerType
 from didiator.interface.exceptions import CommandHandlerNotFound, HandlerNotFound
-from didiator.request_dispatcher import RequestDispatcherImpl
+from didiator.dispatchers.request import RequestDispatcherImpl
 
 CRes = TypeVar("CRes")
 C = TypeVar("C", bound=Command[Any])
