+++ didiator/interface/dispatchers/__init__.py
@@ -1,0 +1,9 @@
+from .command import CommandDispatcher
+from .request import Dispatcher
+from .query import QueryDispatcher
+
+__all__ = (
+    "CommandDispatcher",
+    "Dispatcher",
+    "QueryDispatcher",
+)
