+++ didiator/interface/entities/__init__.py
@@ -1,0 +1,9 @@
+from .command import Command
+from .request import Request
+from .query import Query
+
+__all__ = (
+    "Command",
+    "Request",
+    "Query",
+)
