+++ didiator/interface/entities/request.py
@@ -1,0 +1,8 @@
+import abc
+from typing import Generic, TypeVar
+
+RRes = TypeVar("RRes")
+
+
+class Request(abc.ABC, Generic[RRes]):
+    pass
