+++ didiator/interface/handlers/__init__.py
@@ -1,0 +1,10 @@
+from .command import CommandHandler
+from .request import Handler, HandlerType
+from .query import QueryHandler
+
+__all__ = (
+    "CommandHandler",
+    "Handler",
+    "QueryHandler",
+    "HandlerType",
+)
