+++ didiator/mediator.py
@@ -1,12 +1,12 @@
 from typing import Any, TypeVar
 
-from didiator.command import Command
-from didiator.command_dispatcher import CommandDispatcherImpl
-from didiator.interface.command_dispatcher import CommandDispatcher
+from didiator.dispatchers.command import CommandDispatcherImpl
+from didiator.dispatchers.query import QueryDispatcherImpl
+from didiator.interface.entities.command import Command
+from didiator.interface.dispatchers.command import CommandDispatcher
+from didiator.interface.dispatchers.query import QueryDispatcher
+from didiator.interface.entities.query import Query
 from didiator.interface.mediator import Mediator
-from didiator.interface.query_dispatcher import QueryDispatcher
-from didiator.query import Query
-from didiator.query_dispatcher import QueryDispatcherImpl
 
 CRes = TypeVar("CRes")
 QRes = TypeVar("QRes")
