+++ tests/unit/test_middlewares.py
@@ -1,7 +1,8 @@
 from dataclasses import dataclass
 from functools import partial
 
-from didiator.command import Command, Handler
+from didiator.interface.handlers import CommandHandler
+from didiator.interface.entities.command import Command
 from didiator.middlewares.base import Middleware
 from tests.mocks.middlewares import DataAdderMiddlewareMock, DataRemoverMiddlewareMock
 
@@ -11,7 +12,7 @@ class CommandMock(Command[bool]):
     pass
 
 
-class HandlerMock(Handler[CommandMock, bool]):
+class HandlerMock(CommandHandler[CommandMock, bool]):
     def __init__(self, *args, **kwargs):
         self._excluded_args = args
         self._expected_kwargs = kwargs
